+++ Tests/CleanArchitecture.Comprehensive.HttpClients/CleanArchitecture.Comprehensive.HttpClients.Infrastructure.Stubs/HttpClients/Customers/CustomersServiceHttpClientStub.cs
@@ -1,0 +1,98 @@
+using CleanArchitecture.Comprehensive.HttpClients.Application.IntegrationServices;
+using CleanArchitecture.Comprehensive.HttpClients.Application.IntegrationServices.Contracts.Services.Customers;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: DefaultIntentManaged(Mode.Fully, Targets = Targets.Usings)]
+[assembly: IntentTemplate("Intent.Integration.HttpClients.Stubs.HttpClientStub", Version = "1.0")]
+
+namespace CleanArchitecture.Comprehensive.HttpClients.Infrastructure.Stubs.HttpClients.Customers
+{
+    public class CustomersServiceHttpClientStub : ICustomersService
+    {
+        [IntentManaged(Mode.Fully, Body = Mode.Fully)]
+        public async Task<Guid> CreateCustomerAsync(
+            CreateCustomerCommand command,
+            CancellationToken cancellationToken = default)
+        {
+            return await Task.FromResult(Guid.Empty);
+        }
+
+        [IntentManaged(Mode.Fully, Body = Mode.Fully)]
+        public async Task DeleteCustomerAsync(Guid id, CancellationToken cancellationToken = default)
+        {
+            await Task.CompletedTask;
+        }
+
+        [IntentManaged(Mode.Fully, Body = Mode.Fully)]
+        public async Task<CustomerDto> GetCustomerByIdAsync(Guid id, CancellationToken cancellationToken = default)
+        {
+            return await Task.FromResult(new CustomerDto
+            {
+                Id = Guid.Empty,
+                Email = string.Empty,
+                Name = string.Empty,
+                Surname = string.Empty
+            });
+        }
+
+        [IntentManaged(Mode.Fully, Body = Mode.Fully)]
+        public async Task<List<CustomerDto>> GetCustomerByNameEmailAsync(
+            GetCustomerByNameEmailQuery query,
+            CancellationToken cancellationToken = default)
+        {
+            return await Task.FromResult(new List<CustomerDto>
+            {
+                new CustomerDto
+                {
+                    Id = Guid.Empty,
+                    Email = string.Empty,
+                    Name = string.Empty,
+                    Surname = string.Empty
+                }
+            });
+        }
+
+        [IntentManaged(Mode.Fully, Body = Mode.Fully)]
+        public async Task<List<CustomerDto>> GetCustomerExtraFieldsAsync(
+            GetCustomerExtraFieldsQuery query,
+            CancellationToken cancellationToken = default)
+        {
+            return await Task.FromResult(new List<CustomerDto>
+            {
+                new CustomerDto
+                {
+                    Id = Guid.Empty,
+                    Email = string.Empty,
+                    Name = string.Empty,
+                    Surname = string.Empty
+                }
+            });
+        }
+
+        [IntentManaged(Mode.Fully, Body = Mode.Fully)]
+        public async Task<List<CustomerDto>> GetCustomersAsync(CancellationToken cancellationToken = default)
+        {
+            return await Task.FromResult(new List<CustomerDto>
+            {
+                new CustomerDto
+                {
+                    Id = Guid.Empty,
+                    Email = string.Empty,
+                    Name = string.Empty,
+                    Surname = string.Empty
+                }
+            });
+        }
+
+        [IntentManaged(Mode.Fully, Body = Mode.Fully)]
+        public async Task UpdateCustomerAsync(UpdateCustomerCommand command, CancellationToken cancellationToken = default)
+        {
+            await Task.CompletedTask;
+        }
+
+        public void Dispose()
+        {
+        }
+    }
+}